+++ Assets/Programing/PJM/Scene/SkillTestScene.unity
@@ -4676,7 +4676,6 @@ MonoBehaviour:
   - {fileID: 1167114202}
   - {fileID: 1287562918}
   - {fileID: 223828996}
-  hpBarPrefab: {fileID: 0}
 --- !u!4 &1802699515
 Transform:
   m_ObjectHideFlags: 0
